Add files via upload
This project automates the process of checking your internet speed and tweeting at your internet service provider if the speed falls below the promised rates. Using Selenium WebDriver, the bot measures download and upload speeds from Speedtest.net and posts a complaint on Twitter if the speeds are below the specified thresholds.